Changes for @pabender's code review (Thanks!)

+++ java/src/jmri/jmrit/display/layoutEditor/LayoutTurnout.java
@@ -2001,7 +2001,7 @@ public class LayoutTurnout extends LayoutTrack {
         Point2D p, minPoint = MathUtil.zeroPoint2D;
 
         double circleRadius = LayoutEditor.SIZE * layoutEditor.getTurnoutCircleSize();
-        double distance, minDistance = Float.POSITIVE_INFINITY;
+        double distance, minDistance = Double.POSITIVE_INFINITY;
 
         // check center coordinates
         if (!requireUnconnected) {
